+++ src/lib/js/components/organisms/LocalistView.tsx
@@ -8,7 +8,7 @@ import InlineCompact from "../molecules/InlineCompact";
 import Calendar from "../molecules/Calendar";
 import { EventElement, FilterBy, Format, HideType } from "../../../types/types";
 
-interface Props {
+export interface Props {
   events: EventElement[];
   format: Format;
   truncatedescription: string;
